+++ examples/CHANGELOG.md
@@ -1,5 +1,50 @@
 # performer-examples
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [7fb8f2b]
+  - @performer/playground@0.0.8
+  - @performer/core@0.0.8
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [fd909e4]
+  - @performer/playground@0.0.7
+  - @performer/core@0.0.7
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [9b916f3]
+  - @performer/playground@0.0.6
+  - @performer/core@0.0.6
+
+## 0.0.4
+
+### Patch Changes
+
+- @performer/core@0.0.5
+- @performer/playground@0.0.5
+
+## 0.0.3
+
+### Patch Changes
+
+- 16d7aa7: Exclude examples from publish
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [aaf3a1e]
+  - @performer/core@0.0.4
+  - @performer/playground@0.0.4
+
 ## 0.0.1
 
 ### Patch Changes