+++ packages/create-performer/templates/default/CHANGELOG.md
@@ -4,5 +4,44 @@
 
 ### Patch Changes
 
+- Updated dependencies [7fb8f2b]
+  - @performer/playground@0.0.8
+  - @performer/core@0.0.8
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [fd909e4]
+  - @performer/playground@0.0.7
+  - @performer/core@0.0.7
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [9b916f3]
+  - @performer/playground@0.0.6
+  - @performer/core@0.0.6
+
+## null
+
+### Patch Changes
+
+- @performer/core@0.0.5
+- @performer/playground@0.0.5
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [aaf3a1e]
+  - @performer/core@0.0.4
+  - @performer/playground@0.0.4
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [bdf0e89]
   - @performer/playground@0.0.3