+++ packages/performer-core/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @performer/core
 
+## 0.0.8
+
+## 0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- 9b916f3: Fix playground when used from create template
+
+## 0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- aaf3a1e: Publish without dry run
+
 ## 0.0.2
 
 ### Patch Changes