+++ packages/performer-playground/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @performer/playground
 
+## 0.0.8
+
+### Patch Changes
+
+- 7fb8f2b: Fix playground vite dependency bundling when using playground cli
+  - @performer/core@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- fd909e4: Playground use vite createServer instead of spawn vite
+  - @performer/core@0.0.7
+
+## 0.0.6
+
+### Patch Changes
+
+- 9b916f3: Fix playground when used from create template
+- Updated dependencies [9b916f3]
+  - @performer/core@0.0.6
+
+## 0.0.5
+
+### Patch Changes
+
+- @performer/core@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- aaf3a1e: Publish without dry run
+- Updated dependencies [aaf3a1e]
+  - @performer/core@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes